Add BAT testing document
This commit adds a short document on BAT testing.  This document is an
updated version of the README.md file that was stored in the _release/bat
folder of the ciao repo.  In the interests of avoid document duplication
the README.md file in github.com/ciao-project/ciao/_release/bat will
soon be deleted.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>