Add architecture document
This commit adds an architecture document for CIAO.  The document is
intended to be fairly high level and to act as an introduction to
CIAO, its concepts and terminology.  There's a particular focus on
CIAO's networking model which is one of the most interesting and complicated
parts of the cluster.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>