Update the kubicle docs to reflect a change in ccloudvm
The units of the -mem option to ccloudvm has changed.  Previoiusly GBs were
used.  Now Mebibytes are used.  This change allows users to create VMs that
use less than 1GB of memory which might be useful when using lightweight
workloads.  This commit updates the kubicle documents to reflect the
change.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>